Bugfixes for init/25_debian_backup.sh.

+++ init/25_debian_backup.sh
@@ -4,14 +4,19 @@ resticctl_create_repo_template()
 {
     MY_RESTIC_CFG_PATH=$1
     sudo mkdir -p "$MY_RESTIC_CFG_PATH"
-    MY_PROFILE_NAME=$2
-    MY_REPO_TEMPLATE="$MY_RESTIC_CFG_PATH/${MY_PROFILE_NAME}.repo"
+    MY_REPO_NAME=$2
+    MY_REPO_TEMPLATE="$MY_RESTIC_CFG_PATH/$MY_REPO_NAME.repo"
 
     if [ -f "$MY_REPO_TEMPLATE" ]; then
         return
     fi
 
-    cat > "$MY_REPO_TEMPLATE" <<EOF
+    echo "Enter repository password to use and press [ENTER]: "
+    stty -echo
+    read MY_RESTIC_PASSWORD
+    stty echo
+
+    sudo cat > "$MY_REPO_TEMPLATE" <<EOF
 # Repository configuration file for restic profile '$rname'
 # Created at $(date) by $USER on $(uname -n)
 #
@@ -19,9 +24,14 @@ resticctl_create_repo_template()
 #   1. Don't forget proper quoting where appropriate
 #   2. Shell commands can be used (eg, subshells, variable expansion etc)
 # restic repository and access options
-RESTIC_REPOSITORY=rest:http://192.168.0.120:8000/$MY_PROFILE_NAME
-RESTIC_PASSWORD=changeme
+RESTIC_REPOSITORY=rest:http://192.168.0.120:8000/$MY_REPO_NAME
+RESTIC_PASSWORD=$MY_RESTIC_PASSWORD
 EOF
+
+  # Make sure to clear password.
+  MY_RESTIC_PASSWORD=
+
+  sudo chmod 600 "$MY_REPO_TEMPLATE"
 }
 
 resticctl_create_profile_template()
@@ -29,7 +39,7 @@ resticctl_create_profile_template()
     MY_RESTIC_CFG_PATH=$1
     sudo mkdir -p "$MY_RESTIC_CFG_PATH"
     MY_PROFILE_NAME=$2
-    MY_PROFILE_TEMPLATE="$MY_RESTIC_CFG_PATH/${MY_PROFILE_NAME}.profile"
+    MY_PROFILE_TEMPLATE="$MY_RESTIC_CFG_PATH/$MY_PROFILE_NAME.profile"
 
     if [ -f "$MY_PROFILE_TEMPLATE" ]; then
         return
@@ -75,6 +85,17 @@ KEEP_WEEKLY=5
 KEEP_MONTHLY=12
 KEEP_YEARLY=2
 EOF
+
+    sudo chmod 600 "$MY_PROFILE_TEMPLATE"
+}
+
+resticctl_init_repo()
+{
+    MY_RESTIC_CFG_PATH=$1
+    MY_REPO_NAME=$2
+
+    sudo dev_nas_wol 10
+    sudo resticctl init "$MY_RESTIC_REPO_NAME"
 }
 
 # Install required dependencies.
@@ -84,10 +105,12 @@ sudo apt-get install -y rsync etherwake
 MY_RESTIC_REPO_NAME=$(hostname)
 MY_RESTIC_DIR_BIN=/usr/local/bin
 MY_RESTIC_DIR_TMP=$(mktemp -d)
-MY_RESTIC_VER=0.11.0
-sh -c "$(wget -O "$MY_RESTIC_DIR_TMP/restic.bz2" https://github.com/restic/restic/releases/download/v${MY_RESTIC_VER}/restic_${MY_RESTIC_VER}_linux_amd64.bz2)"
-bunzip2 "$MY_RESTIC_DIR_TMP/restic.bz2"
-sudo install -m 755 "$MY_RESTIC_DIR_TMP/restic" "$MY_RESTIC_DIR_BIN/restic"
+if [ ! -f "$MY_RESTIC_DIR_BIN/restic" ]; then
+    MY_RESTIC_VER=0.11.0
+    sh -c "$(wget -O "$MY_RESTIC_DIR_TMP/restic.bz2" https://github.com/restic/restic/releases/download/v${MY_RESTIC_VER}/restic_${MY_RESTIC_VER}_linux_amd64.bz2)"
+    bunzip2 "$MY_RESTIC_DIR_TMP/restic.bz2"
+    sudo install -m 755 "$MY_RESTIC_DIR_TMP/restic" "$MY_RESTIC_DIR_BIN/restic"
+fi
 sudo restic self-update
 
 # Install restic wrapper script (resticctl, see https://github.com/fukawi2/resticctl).
@@ -101,14 +124,21 @@ sh -c "$(wget -P "$MY_RESTIC_DIR_TMP" https://raw.githubusercontent.com/fukawi2/
 sh -c "$(wget -P "$MY_RESTIC_DIR_TMP" https://raw.githubusercontent.com/fukawi2/resticctl/master/restic%40.timer)"
 sh -c "$(wget -P "$MY_RESTIC_DIR_TMP" https://raw.githubusercontent.com/fukawi2/resticctl/master/restic-cleanup%40.service)"
 sh -c "$(wget -P "$MY_RESTIC_DIR_TMP" https://raw.githubusercontent.com/fukawi2/resticctl/master/restic-cleanup%40.timer)"
-sudo install -m 644 "$MY_RESTIC_DIR_TMP"/restic*.service "$MY_RESTIC_DIR_TMP"/restic*.timer "$MY_SYSTEMD_SYS_DIR"
+sudo install -m 644 "$MY_RESTIC_DIR_TMP/restic*.service" "$MY_RESTIC_DIR_TMP/restic*.timer" "$MY_SYSTEMD_SYS_DIR"
 sudo systemctl enable restic@${MY_RESTIC_REPO_NAME}.timer
 sudo systemctl enable restic-cleanup@${MY_RESTIC_REPO_NAME}.timer
 
 # Initialize repo.
-resticctl_create_repo_template "/etc/restic/" "$MY_RESTIC_REPO_NAME"
-resticctl_create_profile_template "/etc/restic/" "$MY_RESTIC_REPO_NAME"
-sudo resticctl init "$MY_RESTIC_REPO_NAME"
+MY_REPO_DO_INIT=
+sudo resticctl $MY_RESTIC_REPO_NAME
+if [ $? -ne "0" ]; then
+    MY_REPO_DO_INIT=1
+fi
+if [ -n "$MY_REPO_DO_INIT" ]; then
+    resticctl_create_repo_template "/etc/restic/" "$MY_RESTIC_REPO_NAME"
+    resticctl_create_profile_template "/etc/restic/" "$MY_RESTIC_REPO_NAME"
+    resticctl_init_repo "$MY_RESTIC_CFG_PATH" "$MY_RESTIC_REPO_NAME"
+fi
 
 # Cleanup.
 rm -rf "$MY_RESTIC_DIR_TMP"
